requirements.txt
# local package
# -e .
# pytorch-cpu
# tsai

### Data ###
pandas
google-cloud-storage

### ML ###
sklearn

### Viz ###
matplotlib
ipywidgets
plotly.express
jupyter-dash

### Dash App ###
dash-bootstrap-components
redis
fakeredis
pyarrow

### Setup ###
